pyproject.toml
[project]
name = "work-engati"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "certifi>=2024.2.2",
  "pydantic-settings>=2.1",
  "python-dotenv>=1.0",
  "pymongo>=4.6",
  "langchain>=0.2",
  "langchain-community>=0.0.20",
  "langchain-mongodb>=0.1.0",
  "langchain-openai>=0.1.0",
]
